pmp: Support gdb traces without memory address
The awk part of pmp.sh captures the 4th field of lines like this:

    #9  0x0000559d89a8ac1f in mongo::transport::TransportLayerASIO::ASIOSession::read<asio::mutable_buffers_1> (this=this@entry=0x559da3852000, buffers=..., baton=...) at src/mongo/transport/session_asio.h:411

But I also have a lot of lines where the function name is in 2nd field:

    #19 std::_Function_handler<void(), mongo::transport::ServiceExecutorSynchronous::schedule(mongo::transport::ServiceExecutor::Task, mongo::transport::ServiceExecutor::ScheduleFlags, mongo::transport::ServiceExecutorTaskName)::<lambda()> >::_M_invoke(const std::_Any_data &) (__functor=...) at /opt/mongodbtoolchain/v2/include/c++/5.4.0/functional:1871

This commit adds a check to the awk to support both cases.

I addition to not losing function names, this solves a problem where sometimes
the 4th field contains random memory addresses, which would break the
grouping with `uniq`.