Support --delete_per_insert
With this option there is a delete per insert to avoid growing the table. The deletes
are from the oldest entries (by transactionid) of the table. So deletes are from the
tail while inserts are to the head of the table.

This isn't supported yet when only1t (N clients share one table) is set.